CELLO-352 : Handle the WPS return code for RDKProxy & Legacy Plugin (#53)
Reason for change: Below are the changes,
(1) The legacy plugin must return `"result":1` when failed as per the existing documentation. So added this change.
(2) When WPS is initiated and wps failed at the RDKProxy, its returning RPC success but with the return param.status as failure. This has to be handled.
Test Procedure: As per CELLO-352
Risks: Low
Signed-off-by: Karunakaran A <karunakaran_amirthalingam@cable.comcast.com>