DELIA-67124 : Handle the JSON Subscribe Event
DELIA-67124 : Handle the JSON Subscribe Event

Reason for change: The subscription request reached the NetworkManager and responded but due to the boot-up time no-cpu scenario, the success is NOT sent to consumer. So the consumer subscribes for the 2nd time, it returns already_exist error. It should be handled cleanly.
Test Procedure: As per DELIA-67124
Risks: Low