RDKTV-36415 - Device is taking 3-5min to connect the internet after waking from DS (#148)
* RDKTV-36415 - Device is taking 3-5min to connect the internet after waking from DS

Reason for change: To address the scenario of IPv6 lost after the device wakeup from deepsleep
and has only IPv4. Currenlty, we are getting IPv6 as default IP Address which resulted in this
issue. Enhanced the logic to get both IPv4 and IPv6 in GetInitialConnectionState() to address
this issue.
Test Procedure: As mentioned in the jira ticket RDKTV-36415
Risks: Medium
Priority: P0
Signed-off-by: Gururaaja ESR <gururaja_erodesriranganramlingham@comcast.com>

* RDKTV-36415 - Device is taking 3-5min to connect the internet after waking from DS

Reason for change: To address the scenario of IPv6 lost after the device wakeup from deepsleep
and has only IPv4. Currenlty, we are getting IPv6 as default IP Address which resulted in this
issue. Enhanced the logic to get both IPv4 and IPv6 in GetInitialConnectionState() to address
this issue.
Test Procedure: As mentioned in the jira ticket RDKTV-36415
Risks: Medium
Priority: P0
Signed-off-by: Gururaaja ESR <gururaja_erodesriranganramlingham@comcast.com>

* Updated Connectivity Monitoring for more logs

Signed-off-by: kamirt573_comcast <karunakaran_amirthalingam@cable.comcast.com>

---------

Signed-off-by: kamirt573_comcast <karunakaran_amirthalingam@cable.comcast.com>
Co-authored-by: kamirt573_comcast <karunakaran_amirthalingam@cable.comcast.com>