DELIA-68256 - Reset Manual IP is not working (#169)
* DELIA-68256 - Reset Manual IP is not working

Reason for change: Checked for proper ipversion in the input curl
request and it is not valid return as bad request. This will address
the issue of Network Reset performed for AutoConfig IP when the device
is already setup with manual IP, it is not accepting the mode change.
Test Procedure: Check setIPSettings curl command
Risks: Medium
Priority: P2