RDKEMW-4100 - startWiFiSignalQualityMonitor crash
Reason for change: Changed the assignment operator for atomic variable to
load and store to avoid the possibility of '=' operator is overloaded for
atomic types to perform a non-atomic assignment. This means that the
assignment operator would not be thread-safe, even if the variable
is declared as atomic.
Test Procedure: Check startWiFiSignalQualityMonitor works whenever we connect to WiFi
Risks: Medium
Priority: P1