RDKEMW-2444 - nm_event_thrd crash (#177)
Reason for change: The crash is happening because of thread is not joined, when wifi is disconnected while the monitorThreadFunction() is running(which sets the m_isRunning to false) and stopWiFiSignalQualityMonitor() is called. So the next time when the thread creation is initiate it leads to a crash. Removed m_isRunning set to false at the end of monitorThreadFunction(), so that the stopWiFiSignalQualityMonitor() will join the thread in the above mentioned scenario
Test Procedure: Toggle the wifi switch in the UI and check whether nm_event_thrd crash is observed or not
Risks: Medium
Priority: P1