RDKEMW-9249 - Request MfrMgr to Persist WiFi Settings from NetworkManager Plugin (#260)
* RDKEMW-9249 - Request MfrMgr to Persist WiFi Settings from NetworkManager Plugin

Reason for change: To save the wifi credentials to MfrMgr when the wifi connection is successful and remove the same when removeKnownSSID is called
So that when migrating from RDKE to RDKV we will have the wifi configuration to connect. This checkin as the implementation to use gdbus call instead of popen to retrieve the SSID
Test Procedure: CDL upgrade to RDKV from RDKE and check the wifi connection persist
Priority:P1
Risks: Medium
Signed-off-by: Gururaaja ESR<Gururaja_ErodeSriranganRamlingham@comcast.com>

* RDKEMW-9249 - Request MfrMgr to Persist WiFi Settings from NetworkManager Plugin

Reason for change: To save the wifi credentials to MfrMgr when the wifi connection is successful and remove the same when removeKnownSSID is called
So that when migrating from RDKE to RDKV we will have the wifi configuration to connect. This checkin as the implementation to use gdbus call instead of popen to retrieve the SSID
Test Procedure: CDL upgrade to RDKV from RDKE and check the wifi connection persist
Priority:P1
Risks: Medium
Signed-off-by: Gururaaja ESR<Gururaja_ErodeSriranganRamlingham@comcast.com>

---------

Signed-off-by: Gururaaja ESR<Gururaja_ErodeSriranganRamlingham@comcast.com>
Co-authored-by: Karunakaran A <48997923+karuna2git@users.noreply.github.com>